refactor: prepare monorepo for CLI and schematics packages
