+++ .gitignore
@@ -2,7 +2,7 @@
 src/**/*.js
 src/**/*.js.map
 src/**/*.d.ts
-build
+dist
 
 # IDEs
 .idea/
