+++ packages/@guidesmiths/cuckoojs-cli/package.json
@@ -3,20 +3,23 @@
   "version": "0.0.1",
   "description": "CuckooJS CLI",
   "keywords": [
-    "schematics nestjs"
+    "schematics",
+    "nestjs",
+    "cli",
+    "cuckoojs"
   ],
   "license": "MIT",
-  "main": "./build/src/cli/cli.js",
-  "bin": {
-    "cuckoo": "./build/src/cli/cli.js"
+  "engines": {
+    "node": ">= 12"
   },
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "start": "node build/src/cli/cli.js",
-    "test": "npm run build && jasmine src/**/*_spec.js",
-    "lint": "eslint --ext .ts .",
-    "lint:fix": "eslint --fix --ext .ts ."
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/guidesmiths/cuckoojs.git"
   },
+  "bugs": {
+    "url": "https://github.com/guidesmiths/cuckoojs/issues"
+  },
+  "homepage": "https://github.com/guidesmiths/cuckoojs#readme",
   "contributors": [
     {
       "name": "David Yusta",
@@ -35,6 +38,17 @@
       "email": "inigo.marquinez@one-beyond.com"
     }
   ],
+  "main": "dist/src/cli.js",
+  "bin": {
+    "cuckoo": "dist/src/cli.js"
+  },
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "start": "node dist/src/cli.js",
+    "test": "npm run build && jasmine src/**/*_spec.js",
+    "lint": "eslint --ext .ts .",
+    "lint:fix": "eslint --fix --ext .ts ."
+  },
   "dependencies": {
     "@angular-devkit/schematics-cli": "^14.2.3",
     "@guidesmiths/cuckoojs-schematics": "^0.0.1",
