+++ packages/@guidesmiths/cuckoojs-cli/src/commands/abstract.command.ts
