+++ packages/@guidesmiths/cuckoojs-cli/src/commands/generate.command.ts
