+++ packages/@guidesmiths/cuckoojs-cli/src/commands/new.command.ts
