+++ packages/@guidesmiths/cuckoojs-cli/src/lib/runners/bash.runner.husky.ts
