+++ packages/@guidesmiths/cuckoojs-cli/src/lib/runners/schematic.runner.ts
@@ -1,7 +1,7 @@
 import {GenericRunner} from './generic.runner';
 
 export class SchematicRunner extends GenericRunner {
-	private static getSchematicPath(): string {
+	private static getSchematicsCliPath(): string {
 		return require.resolve(
 			'@angular-devkit/schematics-cli/bin/schematics.js',
 			{paths: module.paths},
@@ -12,23 +12,28 @@ export class SchematicRunner extends GenericRunner {
 		super('node');
 	}
 
+	public async listSchematics() {
+		const args = ['@guidesmiths/cuckoojs: --list-schematics'];
+		await super.run({command: SchematicRunner.getSchematicsCliPath(), args, stdio: ['inherit', 'inherit', 'inherit']});
+	}
+
 	public async generateNestApplication(name: string) {
 		const args = ['@nestjs/schematics:application', `--name ${name} --verbosity=QUIET -`];
-		await super.run({command: SchematicRunner.getSchematicPath(), args, stdio: ['pipe', 'pipe', 'inherit']});
+		await super.run({command: SchematicRunner.getSchematicsCliPath(), args, stdio: ['pipe', 'pipe', 'inherit']});
 	}
 
 	public async generateNestElement(schematic: string, name: string, options: string[]) {
 		const args = [`@nestjs/schematics:${schematic} ${name} ${options.join(' ')}`];
-		await super.run({command: SchematicRunner.getSchematicPath(), args, stdio: ['pipe', 'pipe', 'inherit']});
+		await super.run({command: SchematicRunner.getSchematicsCliPath(), args, stdio: ['pipe', 'pipe', 'inherit']});
 	}
 
 	public async addGitignoreFile(name: string) {
 		const args = [`@guidesmiths/cuckoojs-schematics:gitignore --directory=${name}`];
-		await super.run({command: SchematicRunner.getSchematicPath(), args, stdio: ['pipe', 'pipe', 'inherit']});
+		await super.run({command: SchematicRunner.getSchematicsCliPath(), args, stdio: ['pipe', 'pipe', 'inherit']});
 	}
 
 	public async addCommitlint(name: string) {
 		const args = [`@guidesmiths/cuckoojs-schematics:commitlint --directory=${name}`];
-		await super.run({command: SchematicRunner.getSchematicPath(), args, stdio: ['pipe', 'pipe', 'inherit']});
+		await super.run({command: SchematicRunner.getSchematicsCliPath(), args, stdio: ['pipe', 'pipe', 'inherit']});
 	}
 }
