+++ packages/@guidesmiths/cuckoojs-cli/src/lib/ui/ui.ts
@@ -1,4 +1,4 @@
-import {version} from '../../../../package.json';
+import {version} from '../../../package.json';
 
 export const messages: Record<string, string> = {
 	banner: `
