+++ packages/@guidesmiths/cuckoojs-cli/tsconfig.json
@@ -11,7 +11,7 @@
     "noImplicitThis": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "outDir": "build",
+    "outDir": "dist",
     "resolveJsonModule": true,
     "rootDir": ".",
     "skipDefaultLibCheck": true,
@@ -21,6 +21,5 @@
     "target": "es6",
     "types": ["jasmine", "node"]
   },
-  "include": ["src/**/*.ts"],
-  "exclude": ["src/*/files/**/*"]
+  "include": ["src/**/*.ts"]
 }
