+++ packages/@guidesmiths/cuckoojs-schematics/package.json
@@ -2,15 +2,22 @@
   "name": "@guidesmiths/cuckoojs-schematics",
   "version": "0.0.1",
   "description": "A collection of CuckooJS schematics",
-  "scripts": {
-    "build": "tsc -p tsconfig.json",
-    "watch": "tsc -p tsconfig.json --watch",
-    "test": "npm run build && jasmine src/**/*_spec.js"
-  },
   "keywords": [
     "schematics",
     "cuckoojs"
   ],
+  "license": "MIT",
+  "engines": {
+    "node": ">= 12"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/guidesmiths/cuckoojs.git"
+  },
+  "bugs": {
+    "url": "https://github.com/guidesmiths/cuckoojs/issues"
+  },
+  "homepage": "https://github.com/guidesmiths/cuckoojs#readme",
   "contributors": [
     {
       "name": "David Yusta",
@@ -29,8 +36,13 @@
       "email": "inigo.marquinez@one-beyond.com"
     }
   ],
-  "license": "MIT",
-  "schematics": "./src/collection.json",
+  "schematics": "./dist/collection.json",
+  "scripts": {
+    "build": "tsc -p tsconfig.json",
+    "postbuild": "copyfiles -a -u 1 ./src/collection.json ./src/*/schema.json ./src/*/files/** dist/",
+    "watch": "tsc -p tsconfig.json --watch",
+    "test": "npm run build && jasmine src/**/*_spec.js"
+  },
   "dependencies": {
     "@angular-devkit/core": "^14.2.7",
     "@angular-devkit/schematics": "^14.2.6",
@@ -39,6 +51,7 @@
   "devDependencies": {
     "@types/jasmine": "~4.0.0",
     "@types/node": "^14.15.0",
+    "copyfiles": "^2.4.1",
     "jasmine": "^4.0.0"
   }
 }
