--- packages/@guidesmiths/cuckoojs-schematics/src/commitlint/index.d.ts
@@ -1,2 +1,0 @@
-import { Rule } from '@angular-devkit/schematics';
-export declare function run(_options: any): Rule;
