+++ packages/@guidesmiths/cuckoojs-schematics/src/gitignore/files/gitignore
