--- packages/@guidesmiths/cuckoojs-schematics/src/gitignore/gitignore.factory.d.ts
@@ -1,2 +1,0 @@
-import { Rule } from '@angular-devkit/schematics';
-export declare function main(options: any): Rule;
