--- packages/@guidesmiths/cuckoojs-schematics/src/gitignore/gitignore.factory.spec.d.ts
@@ -1,1 +1,0 @@
-export {};
