--- packages/@guidesmiths/cuckoojs-schematics/src/gitignore/index_spec.d.ts
@@ -1,1 +1,0 @@
-export {};
