+++ packages/@guidesmiths/cuckoojs-schematics/tsconfig.json
@@ -1,17 +1,17 @@
 {
   "compilerOptions": {
     "baseUrl": "tsconfig",
-    "lib": ["es2018", "dom"],
     "declaration": true,
+    "lib": ["es2018", "dom"],
     "module": "commonjs",
     "moduleResolution": "node",
     "noEmitOnError": true,
     "noFallthroughCasesInSwitch": true,
     "noImplicitAny": true,
     "noImplicitThis": true,
-    "noUnusedParameters": true,
     "noUnusedLocals": true,
-    "rootDir": "src/",
+    "noUnusedParameters": true,
+    "outDir": "dist",
     "skipDefaultLibCheck": true,
     "skipLibCheck": true,
     "sourceMap": true,
@@ -20,5 +20,5 @@
     "types": ["jasmine", "node"]
   },
   "include": ["src/**/*"],
-  "exclude": ["src/*/files/**/*"]
+  "exclude": ["node_modules", "src/**/*.spec.ts", "src/*/files/**/*"]
 }
